move formatDate to a util
