+++ btr-web/btr-main-app/components/bcros/comments/Comment.vue
@@ -1,17 +1,6 @@
 <script setup lang='ts'>
+import { formatDate } from '~/utils/date-utils'
 defineProps<{ comment: object | string }>()
-
-const formatDate = function (date: string) {
-  const options: Intl.DateTimeFormatOptions = {
-    year: 'numeric',
-    month: 'long',
-    day: 'numeric',
-    hour: 'numeric',
-    minute: 'numeric',
-    timeZoneName: 'long'
-  }
-  return new Date(date).toLocaleDateString('en-US', options)
-}
 </script>
 
 <template>
@@ -24,7 +13,7 @@ const formatDate = function (date: string) {
     <div class="italic mb-6">
       <UIcon name="i-mdi-message-text-outline" class="mr-8 w-6 h-6" />
       <span class="h-6 align-super leading-6">
-        {{ comment.user?.first }} {{ comment.user?.last }} - {{ formatDate(comment.createdAt) }}
+        {{ comment.user?.first }} {{ comment.user?.last }} - {{ formatDate(comment.createdAt, true) }}
       </span>
     </div>
     <p class="mb-2 ml-8 pl-6">
